requirements.txt
alembic==1.13.3
asyncpg==0.30.0
fastapi==0.115.8
httpx==0.28.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
sqlalchemy==2.0.35
uvicorn==0.34.0
psycopg2-binary==2.9.10
